refactor(auth): move auth repository to domain/auth/repositories
Update all imports to use new location and move tests to corresponding directory
