+++ src/presentation/controller/__tests__/controller.test.ts
@@ -4,7 +4,7 @@ import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { UpdateUserDto } from "@/domain/dtos/update-user.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
-import { AuthRepository } from "@/domain/repositories/auth.repository";
+import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
 import { RegisterUser } from "@/domain/auth/use-cases/register-user";
 import { LoginUser } from "@/domain/auth/use-cases/login-user";
 import { UpdateUser } from "@/domain/use-cases/update-user";
